refactor

+++ public/action/account/search_connect_parent.php
@@ -25,7 +25,7 @@ $searchBy = $_POST['searchBy'];
 // Check if empty query
 if(empty($query)){
 
-	$r = $db_parentchild->select(array("parentchild_id"));
+	$r = $db_parentchild->select(array());
 
 } else {
 	
@@ -33,12 +33,12 @@ if(empty($query)){
 
 		case("parent_id"):
 
-			$r = $db_parentchild->like(array("parentchild_id"), "parent_id", "/.*$query/");
+			$r = $db_parentchild->like(array(), "parent_id", "/.*$query/");
 			break;
 
 		case("student_id"):
 
-			$r = $db_parentchild->like(array("parentchild_id"), "student_id", "/.*$query/");
+			$r = $db_parentchild->like(array(), "student_id", "/.*$query/");
 			break;
 
 			}//
@@ -56,52 +56,54 @@ if(empty($r)){
 		</div>
 	</div>";
 
-} else {
+}
+
+if(!empty($r)){
 	
-	// loop along results
-	foreach($r as $key){
-		foreach($key as $parentchild_id){
-            
-			// Get Info
-			$parent_id = $db_parentchild->get("parent_id", "parentchild_id", "$parentchild_id");
-            $student_id = $db_parentchild->get("student_id", "parentchild_id", "$parentchild_id");
+	foreach($r as $pc){
+		$parentchild_id = $pc['parentchild_id'];
+		$parent_id = $pc['parent_id'];
+		$student_id = $pc['student_id'];
+		$relation = $pc['relation'];
 
-            $p_first_name = $db_parent->get("first_name", "parent_id", "$parent_id");
-            $p_last_name = $db_parent->get("last_name", "parent_id", "$parent_id");
-            $p_suffix_name = $db_parent->get("suffix_name", "parent_id", "$parent_id");
-            $p_name = $p_first_name." ".$p_last_name." ".$p_suffix_name;
+		$parent_info = $db_parent->where(array(),"parent_id",$parent_id);
+		$student_info = $db_student->where(array(),"student_id",$student_id);
 
-            $s_first_name = $db_student->get("first_name", "student_id", "$student_id");
-            $s_last_name = $db_student->get("last_name", "student_id", "$student_id");
-            $s_suffix_name = $db_student->get("suffix_name", "student_id", "$student_id");
-            $s_name = $s_first_name." ".$s_last_name." ".$s_suffix_name;
+		foreach($parent_info as $parent){
+			$p_first_name = $parent['first_name'];
+			$p_last_name = $parent['last_name'];
+			$p_suffix_name = $parent['suffix_name'];
+			$p_name = "$p_first_name $p_last_name $p_suffix_name";
+		}
+		foreach($student_info as $student){
+			$s_first_name = $student['first_name'];
+			$s_last_name = $student['last_name'];
+			$s_suffix_name = $student['suffix_name'];
+			$s_name = "$s_first_name $s_last_name $s_suffix_name";
+		}
 
-			// Echo result card
+		echo "
+		<div class='card'>
+			<div class='card-content'>
+				<p><font size='3'><b>Parent: $p_name<br>Child: $s_name ($student_id)</b></font></p>
+		</div>";
+
+		// Show addt'l options for admin							
+		if($_SESSION['account_type'] == "admin"){
+								
 			echo "
-			<div class='card'>
-				<div class='card-content'>
-					<p><font size='3'><b>Parent: $p_name<br>Child: $s_name ($student_id)</b></font></p>
-				</div>";
-
-			// Show addt'l options for admin							
-			if($_SESSION['account_type'] == "admin"){
-									
-				echo "
-				<div class='card-action'>
-					<a class='black-text' href='../../profile/?parent_id=$parent_id'>View Parent</a>
-            		<a class='black-text' href='../../profile/?student_id=$student_id'>View Student</a>
-					<a class='black-text' href='../../action/account/delete_parentchild.php?parentchild_id=$parentchild_id'>Remove Connection</a>
-				</div></div>";
-									
-			}
+			<div class='card-action'>
+				<a class='black-text' href='../../profile/?parent_id=$parent_id'>View Parent</a>
+				<a class='black-text' href='../../profile/?student_id=$student_id'>View Student</a>
+				<a class='black-text' href='../../action/account/delete_parentchild.php?parentchild_id=$parentchild_id'>Remove Connection</a>
+			</div></div>";
+								
+		}
 							
-			// card ender
-			echo "</div>";
-						
-		} // end of inner fe
-				
-	} // end of outer fe
+		// card ender
+		echo "</div>";
+	}
 		
-} // non empty result
+}
 
 ?>
